Refactor: Toast -> 디렉토리 형식으로 변경
